requirements.txt
streamlit
mysql-connector-python
opencv-python
pillow
face_recognition
dlib
numpy
pytz
chromadb

